feat(functions and main) remove login command and fix

+++ bot/functions.py
@@ -6,8 +6,6 @@ import random
 import psycopg2
 from datetime import datetime
 from django.utils.timezone import now
-import html
-import markdown
 
 db_name = "42"
 db_user = "postgres"
@@ -31,41 +29,48 @@ try:
 except (Exception, psycopg2.Error) as error:
     print("PostgreSQL-ga ulanishda xatolik:", error)
 
-async def start_command(message:Message, state:FSMContext):
+async def start_command(message: Message, state:FSMContext):
     user_id = message.from_user.id
-    cursor.execute("SELECT * FROM user_user WHERE telegram_id = %s", (user_id,))
+    cursor.execute("SELECT * FROM user_user WHERE telegram_id = %s", (user_id, ))
     user = cursor.fetchone()
     if user:
+        await message.delete()
         return await send_password(message)
     await message.answer("Kantaktingizni pastdagi tugma orqali kiriting ⬇️", reply_markup=keyboards.contact_markup)
     await state.set_state(states.NewMember.phone)
-
+    await message.delete()
 
 async def get_contact(message:Message, state:FSMContext):
-    cursor.execute("INSERT INTO user_user(password, last_login, telegram_id, full_name, username, phone, created_at, updated_at) VALUES (%s, %s, %s, %s, %s, %s, %s, %s)",
-                   (123, datetime.now(), message.from_user.id, f"{message.from_user.last_name} {message.from_user.first_name}", message.from_user.username, message.contact.phone_number, datetime.now(), datetime.now()))
+    now = datetime.now()
+    last_name = message.from_user.last_name if message.from_user.last_name else "Unknown"
+    first_name = message.from_user.first_name if message.from_user.first_name else "Unknown"
+    username = message.from_user.username if message.from_user.username else "Unknown"
+    phone = message.contact.phone_number
+    id = message.from_user.id
+    cursor.execute("INSERT INTO user_user(last_login, is_superuser, username, first_name, last_name, email, is_staff, is_active, date_joined, password, phone, created_at, updated_at, telegram_id) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)",
+                   (now, False, username, first_name, last_name, None, False, True, now, None, phone, now, now, id))
     connection.commit()
     await state.set_state(states.NewMember.login)
 
-
-async def send_password(message:Message):
-    raqam = random.randint(100000, 999999)
+async def send_password(message: Message):
+    password = random.randint(1000000, 9999999)
     time = datetime.now()
     cursor.execute("SELECT id FROM user_user WHERE telegram_id=%s", (message.from_user.id,))
     user_id = cursor.fetchall()
-    cursor.execute("INSERT INTO user_generatepassword(password, time, user_id) VALUES (%s, %s, %s)", (raqam, time, user_id[0][0]))
+    cursor.execute("INSERT INTO user_generatepassword(password, time, user_id) VALUES (%s, %s, %s)", (password, time, user_id[0][0]))
     connection.commit()
-    await message.answer(str(raqam), reply_markup=keyboards.password_recovery)
+    await message.answer(str(password), reply_markup=keyboards.password_recovery)
 
 
-async def recovery_password(callback_data:CallbackQuery, state:FSMContext):
+async def recovery_password(callback_data:CallbackQuery):
     cursor.execute("SELECT id FROM user_user WHERE telegram_id=%s", (callback_data.from_user.id,))
     user_id = cursor.fetchall()
     cursor.execute("SELECT time FROM user_generatepassword WHERE user_id=%s ORDER BY time DESC", (user_id[0][0],))
     time = cursor.fetchone()
     if (now() - time[0]).total_seconds() < 60:
         return await callback_data.answer("Parolingiz hali kuchda! ☝️")
-    await callback_data.message.delete()
+    if 'message' in callback_data:
+        await callback_data.message.delete()
     raqam = random.randint(100000, 999999)
     time = datetime.now()
     cursor.execute("SELECT * FROM user_user WHERE telegram_id=%s", (callback_data.from_user.id,))
